Modularized the optimization techniques
