+++ src/program/optimizer/util.rs
